Java Project
This is an individual project done to develop an efficient java code having features like code reusability, code security and code flexibility using all the object-oriented principles such as inheritance, polymorphism and abstraction. I this project a simple and efficient java code is developed based on the given problem statement and  got the expected output.